+++ rotkehlchen/data_import/importers/binance.py
@@ -12,7 +12,7 @@ from rotkehlchen.constants import ZERO
 from rotkehlchen.constants.assets import A_USD
 from rotkehlchen.data_import.utils import BaseExchangeImporter, UnsupportedCSVEntry, hash_csv_row
 from rotkehlchen.db.drivers.gevent import DBCursor
-from rotkehlchen.errors.asset import UnknownAsset, UnsupportedAsset
+from rotkehlchen.errors.asset import UnknownAsset
 from rotkehlchen.errors.misc import InputError
 from rotkehlchen.errors.price import NoPriceForGivenTimestamp
 from rotkehlchen.errors.serialization import DeserializationError
@@ -725,7 +725,7 @@ class BinanceImporter(BaseExchangeImporter):
                     is_error=True,
                 )
                 skipped_count += 1
-            except (DeserializationError, UnsupportedAsset) as e:
+            except DeserializationError as e:
                 self.send_message(
                     row_index=index,
                     csv_row=csv_row,