+++ rotkehlchen/db/constants.py
@@ -81,7 +81,6 @@ class UpdateType(Enum):
     ACCOUNTING_RULES = 'accounting_rules'
     LOCATION_ASSET_MAPPINGS = 'location_asset_mappings'
     COUNTERPARTY_ASSET_MAPPINGS = 'counterparty_asset_mappings'
-    LOCATION_UNSUPPORTED_ASSETS = 'location_unsupported_assets'
 
     def serialize(self) -> str:
         """Serializes the update type for the DB and API"""