+++ rotkehlchen/tests/exchanges/test_poloniex.py
@@ -10,7 +10,7 @@ from rotkehlchen.api.websockets.typedefs import WSMessageType
 from rotkehlchen.assets.asset import Asset
 from rotkehlchen.assets.converters import asset_from_poloniex
 from rotkehlchen.constants.assets import A_BCH, A_BTC, A_ETH
-from rotkehlchen.errors.asset import UnknownAsset, UnsupportedAsset
+from rotkehlchen.errors.asset import UnknownAsset
 from rotkehlchen.errors.serialization import DeserializationError
 from rotkehlchen.exchanges.poloniex import Poloniex, trade_from_poloniex
 from rotkehlchen.fval import FVal
@@ -23,14 +23,11 @@ from rotkehlchen.tests.utils.exchanges import (
     POLONIEX_BALANCES_RESPONSE,
     POLONIEX_MOCK_DEPOSIT_WITHDRAWALS_RESPONSE,
     POLONIEX_TRADES_RESPONSE,
-    get_exchange_asset_symbols,
 )
-from rotkehlchen.tests.utils.globaldb import is_asset_symbol_unsupported
 from rotkehlchen.tests.utils.mock import MockResponse
 from rotkehlchen.types import Location, Timestamp, TimestampMS
 
 if TYPE_CHECKING:
-    from rotkehlchen.globaldb.handler import GlobalDBHandler
     from rotkehlchen.tests.fixtures.messages import MockRotkiNotifier
 
 
@@ -312,9 +309,9 @@ def test_query_trade_history_unexpected_data(poloniex):
     given_input = input_trades.replace('"ETH_BTC"', '"0"')
     mock_poloniex_and_query(given_input, expected_warnings_num=0, expected_errors_num=1)
 
-    # symbol with unsupported asset
-    given_input = input_trades.replace('"ETH_BTC"', '"ETH_BALLS"')
-    mock_poloniex_and_query(given_input, expected_warnings_num=1, expected_errors_num=0)
+    # symbol with an asset that can't be mapped to a known asset
+    given_input = input_trades.replace('"ETH_BTC"', '"ETH_NOTAREALASSET"')
+    mock_poloniex_and_query(given_input, expected_warnings_num=0, expected_errors_num=0)
 
     # invalid price
     given_input = input_trades.replace('"0.00003432"', 'null')
@@ -338,17 +335,12 @@ def test_query_trade_history_unexpected_data(poloniex):
 
 
 @pytest.mark.asset_test
-def test_poloniex_assets_are_known(poloniex: 'Poloniex', globaldb: 'GlobalDBHandler'):
-    for asset in get_exchange_asset_symbols(Location.POLONIEX):
-        assert is_asset_symbol_unsupported(globaldb, Location.POLONIEX, asset) is False, f'Poloniex assets {asset} should not be unsupported'  # noqa: E501
-
+def test_poloniex_assets_are_known(poloniex: 'Poloniex'):
     currencies = poloniex.api_query_list('/currencies')
     for asset_data in currencies:
         for poloniex_asset in asset_data:
             try:
                 _ = asset_from_poloniex(poloniex_asset)
-            except UnsupportedAsset:
-                assert is_asset_symbol_unsupported(globaldb, Location.POLONIEX, poloniex_asset)
             except UnknownAsset as e:
                 test_warnings.warn(UserWarning(
                     f'Found unknown asset {e.identifier} with symbol {poloniex_asset} in Poloniex. Support for it has to be added',  # noqa: E501
@@ -358,8 +350,8 @@ def test_poloniex_assets_are_known(poloniex: 'Poloniex', globaldb: 'GlobalDBHand
 @pytest.mark.parametrize('function_scope_initialize_mock_rotki_notifier', [True])
 @pytest.mark.parametrize('use_clean_caching_directory', [True])
 def test_poloniex_query_balances_unknown_asset(poloniex):
-    """Test that if a poloniex balance query returns unknown asset no exception
-    is raised and a warning is generated. Same for unsupported assets"""
+    """Test that if a poloniex balance query returns an asset that can't be mapped
+    no exception is raised and an unknown asset message is generated"""
 
     def mock_unknown_asset_return(url, **kwargs):  # pylint: disable=unused-argument
         return MockResponse(200, POLONIEX_BALANCES_RESPONSE)
@@ -377,15 +369,14 @@ def test_poloniex_query_balances_unknown_asset(poloniex):
 
     messages = poloniex.msg_aggregator.rotki_notifier.messages
     assert len(messages) == 2
-    assert messages[0].message_type == WSMessageType.EXCHANGE_UNKNOWN_ASSET
-    assert 'unsupported poloniex asset CNOTE' in messages[1].data['value']
+    assert all(m.message_type == WSMessageType.EXCHANGE_UNKNOWN_ASSET for m in messages)
 
 
 @pytest.mark.parametrize('function_scope_initialize_mock_rotki_notifier', [True])
 @pytest.mark.parametrize('use_clean_caching_directory', [True])
 def test_poloniex_deposits_withdrawal_unknown_asset(poloniex: 'Poloniex') -> None:
-    """Test that if a poloniex asset movement query returns unknown asset no exception
-    is raised and a warning is generated. Same for unsupported assets"""
+    """Test that if a poloniex asset movement query returns an asset that can't be
+    mapped no exception is raised and an unknown asset message is generated"""
 
     def mock_api_return(url, **kwargs):  # pylint: disable=unused-argument
         if '/trades' in url:
@@ -468,10 +459,7 @@ def test_poloniex_deposits_withdrawal_unknown_asset(poloniex: 'Poloniex') -> Non
 
     messages = cast('MockRotkiNotifier', poloniex.msg_aggregator.rotki_notifier).messages
     assert len(messages) == 4
-    assert messages[0].message_type == WSMessageType.EXCHANGE_UNKNOWN_ASSET
-    assert 'Found withdrawal of unsupported poloniex asset BALLS' in messages[1].data['value']  # type: ignore
-    assert messages[2].message_type == WSMessageType.EXCHANGE_UNKNOWN_ASSET
-    assert 'Found deposit of unsupported poloniex asset EBT' in messages[3].data['value']  # type: ignore
+    assert all(m.message_type == WSMessageType.EXCHANGE_UNKNOWN_ASSET for m in messages)
 
 
 @pytest.mark.parametrize('use_clean_caching_directory', [True])